variable: drop redundant size_t-to-usize casts
size_t is usize, so `length as usize` in the two name scans is a no-op
cast. Clippy (unnecessary_cast) flagged both on PR #40; pass length
directly to from_raw_parts.

+++ src/variable.rs
@@ -321,7 +321,7 @@ unsafe extern "C" fn check_valid_name(
         return;
     }
     // The name is valid unless it contains an unquoted blank or newline.
-    let name_bytes = ::core::slice::from_raw_parts(name as *const u8, length as usize);
+    let name_bytes = ::core::slice::from_raw_parts(name as *const u8, length);
     if !name_bytes
         .iter()
         .any(|&c| stop_set(c, MAP_BLANK | MAP_NEWLINE))
@@ -629,7 +629,7 @@ unsafe extern "C" fn check_variable_reference(name: *const ::core::ffi::c_char, 
         return;
     }
     // The reference is valid unless it contains an unquoted blank or newline.
-    let name_bytes = ::core::slice::from_raw_parts(name as *const u8, length as usize);
+    let name_bytes = ::core::slice::from_raw_parts(name as *const u8, length);
     if !name_bytes
         .iter()
         .any(|&c| stop_set(c, MAP_BLANK | MAP_NEWLINE))
